requirements.txt
torchvision==0.7.0
matplotlib==3.3.1
Flask==1.1.2
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
numpy==1.18.5
Pillow==7.2.0
gunicorn==20.0.4
